+++ test.ts
@@ -1,4 +1,0 @@
-function calculateSum(numbers: number[]) {
-    // First we will check if the numbers array is empty
-
-}